refactor: add providers directory
